+++ NewHorizons/Patches/VolumePatches/FogWarpVolumePatches.cs
@@ -42,29 +42,5 @@ namespace NewHorizons.Patches.VolumePatches
                 return true;
             }
         }
-
-        /// <summary>
-        /// Bug fix from the Outsider
-        /// </summary>
-        /// <returns></returns>
-        [HarmonyPrefix]
-        [HarmonyPatch(typeof(FogWarpVolume), nameof(FogWarpVolume.WarpDetector))]
-        public static bool FogWarpVolume_WarpDetector()
-        {
-            // Do not warp the player if they have entered the fog via a projection
-            return !PlayerState.UsingNomaiRemoteCamera();
-        }
-
-        /// <summary>
-        /// Bug fix from the Outsider
-        /// </summary>
-        /// <returns></returns>
-        [HarmonyPrefix]
-        [HarmonyPatch(typeof(FogWarpDetector), nameof(FogWarpDetector.FixedUpdate))]
-        public static bool FogWarpDetector_FixedUpdate()
-        {
-            // Do not warp the player if they have entered the fog via a projection
-            return !PlayerState.UsingNomaiRemoteCamera();
-        }
     }
 }
